requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.4.0
pymongo==4.8.0
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
email-validator==2.2.0
toml==0.10.2